Cleanup: Remove old modules, transfer tests, fix JSDoc
