+++ lib/buildHelpers/createArchiveMetadata.js
@@ -7,22 +7,25 @@ module.exports = async function(project, buildConfig) {
 	const projectName = project.getName();
 	const type = project.getType();
 
-	let pathMappingSource;
+	const pathMapping = {};
 	switch (type) {
 	case "application":
-		pathMappingSource = "webapp";
+		pathMapping.webapp = `resources/${project.getNamespace()}`;
 		break;
 	case "library":
-	case "legacy-library":
 	case "theme-library":
-		pathMappingSource = "src";
+		pathMapping.src = `resources/${project.getNamespace()}`;
+		pathMapping.test = `test-resources/${project.getNamespace()}`;
+		break;
+	case "legacy-library":
+		pathMapping.src = `resources`;
+		pathMapping.test = `test-resources`;
 		break;
 	default:
 		throw new Error(
 			`Unable to create archive metadata for project ${project.getName()}: ` +
 			`Project type ${type} is currently not supported`);
 	}
-
 
 	const metadata = {
 		specVersion: project.getSpecVersion(),
@@ -47,9 +50,7 @@ module.exports = async function(project, buildConfig) {
 		},
 		resources: {
 			configuration: {
-				paths: {
-					[pathMappingSource]: `resources/${project.getNamespace()}`
-				}
+				paths: pathMapping
 			}
 		}
 	};
