+++ lib/generateProjectGraph.js
@@ -1,7 +1,15 @@
 const path = require("path");
-const projectGraphBuilder = require("./graph/providers/projectGraphBuilder");
+const projectGraphBuilder = require("./graph/projectGraphBuilder");
 const ui5Framework = require("./graph/helpers/ui5Framework");
 const log = require("@ui5/logger").getLogger("generateProjectGraph");
+
+function resolveProjectPaths(cwd, project) {
+	project.path = path.resolve(cwd, project.path);
+	if (project.dependencies) {
+		project.dependencies.forEach((project) => resolveProjectPaths(cwd, project));
+	}
+	return project;
+}
 
 /**
  * Helper module to create a [@ui5/project.graph.ProjectGraph]{@link module:@ui5/project.graph.ProjectGraph}
@@ -51,18 +59,16 @@ const generateProjectGraph = {
 	 *
 	 * @public
 	 * @param {object} options
-	 * @param {object} options.filePath Path to the file dependency configuration file
-	 * @param {string} [options.cwd=process.cwd()] Directory to start searching for the root module
+	 * @param {object} [options.filePath=projectDependencies.yaml] Path to the dependency configuration file
+	 * @param {string} [options.cwd=process.cwd()] Directory to resolve relative paths to
 	 * @param {string} [options.versionOverride] Framework version to use instead of the one defined in the root project
 	 * @returns {Promise<module:@ui5/project.graph.ProjectGraph>} Promise resolving to a Project Graph instance
 	 */
-	usingStaticFile: async function({cwd, filePath, versionOverride}) {
+	usingStaticFile: async function({cwd, filePath = "projectDependencies.yaml", versionOverride}) {
 		log.verbose(`Creating project graph using static file...`);
-		const staticTranslator = require("./translators/static");
 
-		const dependencyTree = await staticTranslator.generateDependencyTree(cwd ? path.resolve(cwd) : process.cwd(), {
-			parameters: [filePath] // *sigh*
-		});
+		const dependencyTree = await generateProjectGraph
+			._readDependencyConfigFile(cwd ? path.resolve(cwd) : process.cwd(), filePath);
 
 		const DependencyTreeProvider = require("./graph/providers/DependencyTree");
 		const provider = new DependencyTreeProvider({
@@ -100,6 +106,30 @@ const generateProjectGraph = {
 		await ui5Framework.enrichProjectGraph(projectGraph, {versionOverride});
 
 		return projectGraph;
+	},
+
+	async _readDependencyConfigFile(cwd, filePath) {
+		const fs = require("graceful-fs");
+		const {promisify} = require("util");
+		const readFile = promisify(fs.readFile);
+		const parseYaml = require("js-yaml").load;
+
+		if (!path.isAbsolute(filePath)) {
+			filePath = path.join(cwd, filePath);
+		}
+
+		let dependencyTree;
+		try {
+			const contents = await readFile(filePath, {encoding: "utf-8"});
+			dependencyTree = parseYaml(contents, {
+				filename: filePath
+			});
+		} catch (err) {
+			throw new Error(
+				`Failed to load dependency tree configuration from path ${filePath}: ${err.message}`);
+		}
+		resolveProjectPaths(cwd, dependencyTree);
+		return dependencyTree;
 	}
 };
