+++ lib/graph/helpers/ui5Framework.js
@@ -131,14 +131,15 @@ module.exports = {
 	 * @param {object} [options]
 	 * @param {string} [options.versionOverride] Framework version to use instead of the root projects framework
 	 *   version from the provided <code>tree</code>
-	 * @returns {Promise<object|null>} Promise
+	 * @returns {Promise<module:@ui5/project.graph.ProjectGraph>}
+	 *   Promise resolving with the given graph instance to allow method chaining
 	 */
 	enrichProjectGraph: async function(projectGraph, options = {}) {
 		const rootProject = projectGraph.getRoot();
 
 		if (rootProject.isFrameworkProject()) {
 			// Ignoring UI5 Framework libraries in dependencies
-			return;
+			return projectGraph;
 		}
 
 		const frameworkName = rootProject.getFrameworkName();
@@ -182,7 +183,7 @@ module.exports = {
 			log.verbose(
 				`No ${frameworkName} libraries referenced in project ${rootProject.getName()} ` +
 				`or in any of its dependencies`);
-			return null;
+			return projectGraph;
 		}
 
 		log.info(`Using ${frameworkName} version: ${version}`);
