+++ lib/graph/projectGraphBuilder.js
@@ -1,8 +1,8 @@
 const path = require("path");
-const Module = require("../Module");
-const ProjectGraph = require("../ProjectGraph");
-const ShimCollection = require("../ShimCollection");
-const log = require("@ui5/logger").getLogger("graph:providers:projectGraphBuilder");
+const Module = require("./Module");
+const ProjectGraph = require("./ProjectGraph");
+const ShimCollection = require("./ShimCollection");
+const log = require("@ui5/logger").getLogger("graph:projectGraphBuilder");
 
 function _handleExtensions(graph, shimCollection, extensions) {
 	extensions.forEach((extension) => {
