+++ lib/specifications/ComponentProject.js
@@ -27,14 +27,14 @@ class ComponentProject extends Project {
 	}
 
 	/**
-	* @public
+	* @private
 	*/
 	getCopyright() {
 		return this._config.metadata.copyright;
 	}
 
 	/**
-	* @public
+	* @private
 	*/
 	getComponentPreloadPaths() {
 		return this._config.builder && this._config.builder.componentPreload &&
@@ -42,7 +42,7 @@ class ComponentProject extends Project {
 	}
 
 	/**
-	* @public
+	* @private
 	*/
 	getComponentPreloadNamespaces() {
 		return this._config.builder && this._config.builder.componentPreload &&
@@ -50,7 +50,7 @@ class ComponentProject extends Project {
 	}
 
 	/**
-	* @public
+	* @private
 	*/
 	getComponentPreloadExcludes() {
 		return this._config.builder && this._config.builder.componentPreload &&
@@ -58,14 +58,14 @@ class ComponentProject extends Project {
 	}
 
 	/**
-	* @public
+	* @private
 	*/
 	getJsdocExcludes() {
 		return this._config.builder && this._config.builder.jsdoc && this._config.builder.jsdoc.excludes || [];
 	}
 
 	/**
-	* @public
+	* @private
 	*/
 	getMinificationExcludes() {
 		return this._config.builder && this._config.builder.minification &&
@@ -73,14 +73,14 @@ class ComponentProject extends Project {
 	}
 
 	/**
-	* @public
+	* @private
 	*/
 	getBundles() {
 		return this._config.builder && this._config.builder.bundles || [];
 	}
 
 	/**
-	* @public
+	* @private
 	*/
 	getPropertiesFileSourceEncoding() {
 		return this._config.resources && this._config.resources.configuration &&
