+++ lib/specifications/Project.js
@@ -29,23 +29,23 @@ class Project extends Specification {
 	}
 
 	/**
-	* @public
+	* @private
 	*/
 	getFrameworkName() {
-		return this._config.framework && this._config.framework.name;
+		return this._config.framework?.name;
 	}
 	/**
-	* @public
+	* @private
 	*/
 	getFrameworkVersion() {
-		return this._config.framework && this._config.framework.version;
+		return this._config.framework?.version;
 	}
 	/**
-	* @public
+	* @private
 	*/
 	getFrameworkDependencies() {
 		// TODO: Clone or freeze object before exposing?
-		return this._config.framework && this._config.framework.libraries || [];
+		return this._config.framework?.libraries || [];
 	}
 
 	isFrameworkProject() {
@@ -57,19 +57,19 @@ class Project extends Specification {
 	}
 
 	getBuilderResourceExcludes() {
-		return this._config.builder && this._config.builder.resources && this._config.builder.resources.excludes || [];
+		return this._config.builder?.resources?.excludes || [];
 	}
 
 	getCustomTasks() {
-		return this._config.builder && this._config.builder.customTasks || [];
+		return this._config.builder?.customTasks || [];
 	}
 
 	getServerSettings() {
-		return this._config.server && this._config.server.settings;
+		return this._config.server?.settings;
 	}
 
 	getBuilderSettings() {
-		return this._config.builder && this._config.builder.settings;
+		return this._config.builder?.settings;
 	}
 
 	getArchiveMetadata() {
