+++ lib/specifications/types/LegacyLibrary.js
@@ -52,6 +52,12 @@ class LegacyLibrary extends Library {
 		return testReader;
 	}
 
+	/**
+	 * Legacy libraries have resources outside their namespace or multiple namespaces
+	 * Therefore it is necessary to remove the namespace form any virtual base paths
+	 *
+	 * @param {string} string Virtual base path to remove an eventual namespace from
+	 */
 	_stripNamespace(string) {
 		return string.replace(this._namespace + "/", "");
 	}
