+++ lib/specifications/types/Library.js
@@ -20,7 +20,8 @@ class Library extends ComponentProject {
 
 	/* === Attributes === */
 	/**
-	* @public
+	*
+	* @private
 	*/
 	getLibraryPreloadExcludes() {
 		return this._config.builder && this._config.builder.libraryPreload &&
@@ -28,12 +29,12 @@ class Library extends ComponentProject {
 	}
 
 	/* === Resource Access === */
-	/*
+
+	/**
 	 *
 	 * Get a resource reader for the sources of the project (excluding any test resources)
 	 * In the future the path structure can be flat or namespaced depending on the project
 	 *
-	 * @public
 	 * @returns {module:@ui5/fs.ReaderCollection} Reader collection
 	*/
 	_getSourceReader() {
