+++ lib/specifications/types/ThemeLibrary.js
@@ -13,7 +13,7 @@ class ThemeLibrary extends Project {
 
 	/* === Attributes === */
 	/**
-	* @public
+	* @private
 	*/
 	getCopyright() {
 		return this._config.metadata.copyright;
